+++ models/Repository/Product.php
@@ -103,7 +103,7 @@ class Product extends \yii\db\ActiveRecord
 
         $unit           = new Unit($data['unit']);
         $product->name  = trim($data['name']);
-        $product->count = (int) $product->count + $unit->convert((float) $data['count']);
+        $product->count = (float) $product->count + $unit->convert((float) $data['count']);
         $product->unit  = $unit->getMainUnit();
 
         return $product;
